requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
transformer
torch
nltk
libretranslatepy
feedparser
openai
google-generativeai